Stable Release 10.0.0
There is a known regression since the last release (#2490), for which the next
release would normally be held. But it is unknown how long it will take to
resolve this issue and there are downstream users wanting features that will
land in 10.0.0 (#2194). This change makes the new release, under the reasoning
that this is the path most helpful to users.

Gitlab: #2194, #2490, closes #2500