core plugin: fix font sizes being limited to 1 in PIC output format
This mistake in 0373b5a9f7fc9a3a8501f746e336b0dcc69b6367 resulted in node and
edge labels effectively being invisible in PIC output. With this fix, there is
still a fairly show-stopping issue with the PIC output format in that node
shapes are drawn in the wrong locations.

Gitlab: #2487