xdot sprintXDot: remove stack buffer backing 'xb'
User-derived data is written to `xb`. There is no reason to believe this data
will be ≤ `BUFSIZ`. So removing this and letting the agxbuf default logic apply
is more straightforward. The final call to `agxbdisown` also forces the buffer
to be heap allocated, so this change simply makes this heap allocation happen
earlier.