common poly_gencode: undo caching of 'AF' allocation
This kind of micro-optimization is unnecessary on contemporary machines. It is
also a barrier to introducing multi-threading.

It is unclear to me what the motivation for the `+ 5` factor is here. The
callees of this function access beyond `size` entries in this array, but their
greatest access is `AF[3]` or `AF[sides + 1]` depending on the value of `sides`.
So the factor seems like it should be `+ 4`. But this is left as-is.