common: zero initialize some 'inside_t' objects
This will become relevant when an upcoming change adds members that need to have
initial values of 0. This change conservatively applies this to every location
an `inside_t` is being constructed with the `s` member in use, though some of
them may not strictly need these new members to be zeroed.

Gitlab: #2498