API BREAK: move 'point_inside' state into 'inside_t'
2c4b2a53b0ea16b004c5a15cd11aadee02caf844 was the result of a discovery that
`poly_inside` uses uninitialized pointers. From this, there is a reasonable
suspicion that the same misbehavior is possible in all `*_inside` functions.
While there is no straightforward way to catch this in the case of
`point_inside` (misbehavior will result in incorrectly comparing against a
radius of 0), it seems wise to evict its state to be caller-managed instead of
living in `static` variables. This also removes a blocker to introducing
multi-threading.

Gitlab: #2498
Suggested-by: Stephen North <scnorth@gmail.com>
Suggested-by: Zhi An Ng <zhin@google.com>