gvc gvprintnum: simplify, using '%.03f' and 'agxbuf_trim_zeros'
There was no need to spell out this error prone logic when we already have what
we need implemented in `agxbuf_trim_zeros` and the standard library. This was
validated by copying the `GVPRINTNUM_TEST`-guarded code into a separate file (it
no longer compiles where it is) and making the changes in the test program to
confirm they have no functional impact on the test cases there.