ccomps cmp: rephrase comparator to avoid signed subtraction
Signed overflow is Undefined Behavior (UB) with respect to the C standard. It
has recently emerged that not only can subtraction of signed values within a
`qsort` comparator like this cause UB, it can also cause out-of-bounds accesses
within the C standard library.¹ While overflow should be impractical here (it
should be impossible to create a graph with a negative number of nodes, let
alone a number of nodes close to `INT_MAX`), it seems wise to rephrase this into
more obviously safe code.

Seems this was missed in the series merged in
4d89a606c58025d1da1c1d52b32cb5503d82d451.

¹ https://seclists.org/oss-sec/2024/q1/69