CI: remove Cygwin “Autotools using autogen.sh” release artifact
During the last release attempt, it was observed packaging failed because a soft
limit of 50 release artifacts has been exceeded. Included in the 52 current
artifacts are various Cygwin files:

  1. Cygwin Autotools package
  2. Cygwin Autotools using autogen.sh package
  3. Cygwin CMake package (.tar.bz and .zip)

It seems reasonable to assume that the commit that added 2
(53a9e5ea1040125cb0304f15b9d472c482142904) intended to add this purely for
testing, not for release.

Gitlab: #2500