CI: remove Cygwin CMake .zip release artifact
During the last release attempt, it was observed packaging failed because a soft
limit of 50 release artifacts has been exceeded. Included in the 52 current
artifacts are various Cygwin files:

  1. Cygwin Autotools package
  2. Cygwin Autotools using autogen.sh package
  3. Cygwin CMake package (.tar.bz and .zip)

In (3), having two separate archive formats for a non-mainstream platform seems
excessive. This change drops it to one, in the hopes of returning to a
releasable state.

Gitlab: #2500