CI: run most of the release deployment flow on every commit
The deployment script was structured to bail out if not on an actual release
commit. A side effect of this is that release problems are only discovered when
actually trying to cut a release. By running everything except the artifact
upload and release creation on every commit, hopefully problems should be caught
in advance of a release itself.

Gitlab: #2500