sfio sfvscanf: remove format stack support
Similar to the prior commit removing this support from `sfprintf`, the code does
not use the generality of this stacking support.

This change removes this stack support to reduce the amount of ongoing code to
maintain. No functionality is affected. The “pop” operation that was previously
triggered and would discover an empty stack, now unconditionally ends `sfvscanf`
as it was naturally doing already.