sfio sfprint: remove format stack support
SFIO’s print function includes an elegant mechanism lacking from libc’s stdio
`printf`: the ability to supply a custom callback for how to handle format
directives. On top of this, it allows the directive `%!` to “stack” further
callbacks. This means, with a single call into `sfprintf`, you can describe how
to format the output while altering the way formatting directives are
interpreted mid-stream.

While this generality is quite creative, it has never been used in Graphviz. The
code uses `%!` in exactly one instance: to pass a format string from GVPR into
SFIO.

This change removes this stack support to reduce the amount of ongoing code to
maintain. No functionality is affected. The “pop” operation that was previously
triggered and would discover an empty stack, now unconditionally ends `sfprintf`
as it was naturally doing already.