gvpr: use standard system allocator for target name
Vmalloc is an arena allocator¹ providing functions `vmalloc` and `vmfree`. Its
main utility in Graphviz is avoiding the need to precisely track memory
allocations for a GVPR program, instead freeing everything on deallocation of
the program itself. There is no need to rely on this for the `tgtname` state
member, whose lifetime is bounded and well understood.

¹ https://en.wikipedia.org/wiki/Region-based_memory_management