common canontoken: fix: remove static buffer
This removes another obstacle to multi-threading. It also removes the puzzling
`+ 10` in the allocation. It is unclear to me what the purpose of this extra
room added in ede18e900076a377908b7daade95ee4883e6437b is. The data that is then
written to this buffer is known to be exactly `len + 1` bytes.

This also removes a bizarre foot gun wherein libgvc’s `gvrender_resolve_color`
is also calling `canontoken`. Because `canontoken` used a static buffer to
return its result, `resolveColor` and `gvrender_resolve_color` could stomp each
other’s cached results if calls to them were interleaved. This could lead to
both false positives (one of these functions believes it already has the color
it needs but it has actually cached a completely different color) and false
negatives (one of these functions believe it needs to re-search, but it actually
already has the target color cached).

Note that this required ensuring the return value of `resolveColor` is always on
the heap to simplify logic in its caller.

This change does nothing about the orthogonal discrepancy that `colorxlate`
performs its color name comparisons case insensitively while
`gvrender_resolve_color` performs its case sensitively.