common colxlate: avoid unnecessary re-search of differently cased colors
fa99ac88ebc6746072a6975eff8ad4e76796a42f altered the color search to be case
insensitive, but it left this guard as case sensitive. As a result, querying a
color under a differently cased name would incorrectly determine it needed to
search for the color, only to re-discover the same cached color it already had.