gc: use central cluster detection function
Prefixing your graph’s name with “cluster” used to be the only way to indicate
clustering. f63e51ab56a6927313524b31b335e9801ff6b09a that added this detection
method seems to have decided this was to be checked case-insensitively. While
this makes sense as the Graphviz keywords are treated case-insensitively, this
seems to disagree with all the other cluster checking in Graphviz source code at
the time which checked case-_sensitively_. This is relevant because the graph
name, which is the thing being checked here, is parsed case-sensitively.

Later, ae160d36572253aa4edfdb14fefe2729bbdfc771 added a second way of indicating
clustering through an attribute. The majority of Graphviz code, like that
touched in this commit, does not use this function and continues to hand roll
its own cluster detection. So most code misses clusters that are specified in
the input file. This is a step towards addressing the discrepancy and aligning
all tools on “is this a cluster?” logic.

Gitlab: #1399