neatogen makeSelfArcs: silence -Wstringop-overflow warning
When building using CMake and `-DCMAKE_BUILD_TYPE=Release` on MinGW, the
compiler says:

  C:/GitLab-Runner/builds/smattr/graphviz/lib/neatogen/neatosplines.c: In function 'makeSelfArcs':
  C:/GitLab-Runner/builds/smattr/graphviz/lib/neatogen/neatosplines.c:242:9: error: 'makeSelfEdge' accessing 8 bytes in a region of size 0 [-Werror=stringop-overflow=]
    242 |         makeSelfEdge(edges, 0, cnt, stepx, stepx, &sinfo);
        |         ^
  C:/GitLab-Runner/builds/smattr/graphviz/lib/neatogen/neatosplines.c:242:9: note: referencing argument 1 of type 'struct edge_t *[0:]'
  C:/GitLab-Runner/builds/smattr/graphviz/lib/common/render.h:112:21: note: in a call to function 'makeSelfEdge'
    112 |     RENDER_API void makeSelfEdge(edge_t * edges[], int ind, int cnt,
        |                     ^

I believe this is a false positive. It does not seem possible to invoke a
function on edge whose edge count you then find to be 0. But this change means
we can avoid thinking about it and get closer to a warning-free release build.