common: more tightly scope 'polypoints' allocation
The values in this array do not need to be retained across calls to the
functions in which they are allocated. Retaining them is a deoptimization in
contemporary environments.

This change also migrates to the newer allocation wrappers. The
lib/cgraph/alloc.h wrappers are similar to the older lib/common/memory.h
wrappers except (1) they are header-only and (2) they live in a directory
(cgraph) that is at the root of the dependency tree. The long term plan is to
replace all use of lib/common/memory.h with lib/cgraph/alloc.h.