core plugin: warn when producing a PS file with an invalid font name
This is not ideal, but there does not seem to be a better alternative. The
infamous font name document referenced in the comment in this change, which
apparently is authoritative, gives no way of escaping characters. Avoiding this
would require proper font name translation which is beyond the goals of the core
plugin.

Gitlab: closes #218
Reported-by: wvh