common point_init: fix out of bounds memory write
This bug was introduced in beede1476b3b23cd22ca1cb0f3b089aa397e42ab. From the
diff, it appears that the extra points were only intended to be appended in the
case when the earlier branch was entered. Adding points that are relative to the
periphery does not seem to make sense when there is no periphery.

Gitlab: fixes #2497
Reported-by: S Scorpionfly