expr: remove match function parameterization
lib/expr appears to have been designed in a highly generic way, suitable for
external consumers. This function pointer allowed callers to describe a custom
string comparison algorithm. lib/expr’s current use is entirely internal to
Graphviz; it is not shipped to users. Graphviz always set this custom string
comparison, but to a function (`matchval`) that does an identical operation to
the default (`strmatch`). This change removes this parameterization to simplify
ongoing maintenance.