neatogen move_node: make 'a' non-static
This allocation does not need to survive across calls to `move_node`. This is a
micro-optimization that is no longer necessary in contemporary environments.
Making this non-static removes a barrier to using multi-threading in future.

This also migrates usage of `ALLOC` to `gv_calloc`. The lib/cgraph/alloc.h
wrappers are similar to the older lib/common/memory.h wrappers except (1) they
are header-only and (2) they live in a directory (cgraph) that is at the root of
the dependency tree. The long term plan is to replace all use of
lib/common/memory.h with lib/cgraph/alloc.h.

Given `Ndim ≤ MAXDIM`, it may be worth dropping dynamic allocation here
altogether in future. Allocating `MAXDIM² × sizeof(double)`
(`100 × sizeof(double)`) on the stack is likely fine on all currently supported
platforms.