split rank crashing test case, reverting prior minimization
This effectively reverts b9e88ae6fa86758d0c358124857058094f5df05e. While that
commit minimized the crash, it minimized the example to something that is
expected to work rather than be rejected. I think it is worthwhile retaining
both the original malformed test case and the minimized expected-to-pass test
case.

Gitlab: #1308