dotgen: fix tracking of number of ranks during rank installation
The `n` and `an` members are intended to track the “number of nodes” and
“allocated number of nodes”, respectively, in a rank. That makes this fix look
straightforward, as it is simply aligning the values here with the next line
performing the actual allocation. However cross-referencing this with another
allocation site of ranking data in lib/dotgen/flat.c suggests there may be
expected to be a +2 fudge factor here. Though the latter code lies in a function
called `abomination`, so perhaps we should not take this as an educational
usage.

I mention the above to explain that I am not sure this is the correct fix. But
it addresses the problems of #1308 and does not break anything else in the test
suite.

Gitlab: fixes #1308
Reported-by: Google Autofuzz project