dotgen abomination: use cgraph wrapper for allocation
The lib/cgraph/alloc.h wrappers are similar to the older lib/common/memory.h
wrappers except (1) they are header-only and (2) they live in a directory
(cgraph) that is at the root of the dependency tree. The long term plan is to
replace all use of lib/common/memory.h with lib/cgraph/alloc.h.

The surrounding code here looks pretty questionable. In particular,
`GD_rank(g) = rptr + 1` appears to lose the reference to the allocated memory
in a way that is not recovered later. Perhaps the title of the containing
function is accurate.