dotgen dot_splines_: use cgraph wrapper for allocation
The lib/cgraph/alloc.h wrappers are similar to the older lib/common/memory.h
wrappers except (1) they are header-only and (2) they live in a directory
(cgraph) that is at the root of the dependency tree. The long term plan is to
replace all use of lib/common/memory.h with lib/cgraph/alloc.h.

An attempt was made to remove manual memory management here altogether and
instead use the generic list implementation (lib/cgraph/list.h). However it
seems this code relies on accessing elements beyond `n_edges` in a way that is
not immediately clear.

This change also updates `GROWEDGES` to use a more standard macro idiom for
statements.