common: treat 'distortion' comparisons against zero as precise
This is a rethink of part of 3611491567a0c4862951a787caa23b32509833b5.
Distortion values are read from the user attribute `distortion` and used as
input values to computations but never themselves the output of a computation.
That is, no floating-point arithmetic imprecision accumulates in distortion
values. So we can be more explicit about our tests of them, squashing 4
-Wfloat-equal compiler warnings.