common isBox: use more nuanced, simpler determination of 'orientation % 90 == 0'
This is a rethink of bfa659ab7125e1e56e4e6108630f52b33bafe1b4. While the
motivation of avoiding truncation to `int` is valid, it seems wiser to take a
more intent-aware approach to what is being computed here. That is, to simply
avoid rounding altogether and achieve this via the comparison itself. It is not
clear `DBL_EPSILON` is an appropriate tolerance to use in the `is_zero`
function.