Autotools: remove commented out -Wdouble-promotion enabling
An initial attempt was made to enable this in
490481afaf4fffa92afbf73716f89b90da30b040, but this was subsequently commented
out due to lacking compiler support. There has been an effort to incrementally
re-enable the commented out warnings (e.g.
f54fdcd898c2a8779abec27e2f4e544f012f60ca). However, in 2024 it no longer seems
desirable to enable -Wdouble-promotion. Double-precision floating-point support
is near-ubiquitous in hardware and there are numerous places where Graphviz code
implicitly promotes `float` values to `double`. If any of these are shown to be
a performance concern, we could revisit this, but for the most part Graphviz
code assumes `double` is supported and never a relevant bottleneck for any of
its operations.