dotgen dot_splines_: use an 'unsigned' for 'cnt'
Squashes a -Wsign-conversion warning. This introduces some cast noise, but this
will be paid down in future commits. Note that `unsigned` was chosen here rather
than `size_t` as we expect this to be a small-ish value and want something that
can be coerced into a `double` without loss of precision (will become relevant
in a future commit).