common: index 'STheap' with 'size_t' values
This is another step towards pervasively using `size_t` for counting things.
Note that, as discussed in a prior commit, the only valid values for
`heap_index` were `-1` and non-negative values. So this change merely swaps in
`SIZE_MAX` for the sentinel value instead of `-1`.

This change expands the maximum supported heap capacity from `INT_MAX` to
`SIZE_MAX - 1` (on x86-64, 2³¹ - 1 → 2⁶⁴ - 1). But this is not currently known
to be a limiting factor and it is assumed this extra flexibility is irrelevant.