add a test case for fdp, neato non-termination when peripheries=0
This triggers the same problem as the previous test case, but for
another reason, even though the root cause is the same. This version
of the test is not affected by the change in
b1a03d58771a11977cd46fdceed6cd3227e512d6 and triggers the problem even
prior to this commit. See the test file for details. The rounding
mentioned in the test file was introduced in
a865bb66aef7e751f2c1a235819c77007b16050f.

This only reproduces on platforms where GTS is built.

GitLab: #2168