cgraph: remove use of '\f5' in man page
This is supposed to select the font mounted in position 5.¹ However it seems to
enable underline that is then not correctly turned off. This is the only section
of the man page that uses `\f5`, so this change switches to using `\fI` like the
rest of the man page instead.

¹ https://www.gnu.org/software/groff/manual/groff.html#index-_005cf