add a '--filepath=…' option to replace '$GV_FILE_PATH'
Many of the use cases served by the previously removed `$GV_FILE_PATH` could be
replicated with `-Gimagepath=…`. Unfortunately there were some that could not.
In particular, using a relative path reference beginning with a slash (e.g.
`image="/foo.png"`).

This change implements a replacement that is intended to cover all previous use
cases. Instead of following the design discussed on the forum,¹ this instead
essentially reverts most of 1e45fd838a7308f0a8e9712d8dbc2e486c265455. The new
mechanism thus works more or less identically to the old one with the exceptions
of (1) being controlled via a command line option instead of an environment
variable and (2) not being intertwined with the `$SERVER_NAME` environment
variable.

This mechanism is not intended as a security or sandboxing technology. If you
are accepting untrusted input and processing it with Graphviz, it is assumed you
are running Graphviz within some other containment technology that restricts its
actions.

Gitlab: closes #2396
Reported-by: Lannar Dean

¹ https://forum.graphviz.org/t/removal-of-gv-file-path-environment-variable