gvc gv_graph_state: remove mistaken use of 'GVATTR_STRING'
Passing `GVATTR_STRING` (0) here makes no sense. It adds a null pointer to
`gv_argvlist_t.argv`, which is a reasonable thing to do, but using the unrelated
`GVATTR_STRING` constant is confusing. The commit message from the change that
added this (392fa4ee765a616985a1383e40d6794f6415442b) suggests maybe this was a
mistake or something planned for that never worked out.