makeObstacle: fix too small edge routing obstacles for ellipse shaped nodes
Create the obstacle by circumscribing an octagon about the ellipse,
instead of inscribing it. Before this, a compass point could be
located outside the obstacle. This caused `raySeg`, called from
`addEndpoint` in `multisplince.c`, to not find any obstacle segment
intersection, when tracing a ray from the compass point outwards from
the node. This had the effect that no edge was added, which triggered
an assertion.

Fixes https://gitlab.com/graphviz/graphviz/-/issues/2168.