circogen: maintain block count external to 'circularLayout'
`initGraphAttrs` attempted to only zero the block count if the root of the graph
being currently dealt with differed from the root of the previous graph being
dealt with. This was subject to both false positives and false negatives:

  1. If a graph was deallocated and then a new graph allocated that happened to
     be placed at the same heap address, the block count of the old graph would
     incorrectly be reused.

  2. If subgraphs A, B, and C were processed, A and C had the same root graph
     but B did not, the block count after processing A would not be reused when
     processing C.

To avoid both of these scenarios, this change moves the owning location of this
count to outside `circularLayout`, where the caller (who knows when reuse should
happen) can simply indicate whether the block count should be reused or not by
the value they pass in.