CMake: standardize on explicit 'PUBLIC'/'PRIVATE' for linkage
This has little effect for executables, but causes some important differences
for libraries. It is better to specify these explicitly because, when a new
`target_link_libraries` statement tries to use `PUBLIC` or `PRIVATE`, CMake
demands that every other `target_link_libraries` statement for the same target
must also be explicit. This change thus reduces the diff for future changes.

Gitlab: !3702