CI: add Fedora 40 jobs
This puts the number of release artifacts even further (54) over the limit (50).
But it seems preferable to have coverage on Fedora 40 than to prioritize paying
down this excess right now. When CentOS 7 and Fedora 38 go EOL soon, dropping
them should bring us back to 50.

This change also disables the PHP bindings on Fedora ≥ 40 because they do not
seem to build correctly:

  error: File not found: /root/rpmbuild/BUILDROOT/graphviz-11.0.1~dev.20240502.0729-1.fc40.x86_64/usr/lib64/php/modules/gv.so
  error: File not found: /root/rpmbuild/BUILDROOT/graphviz-11.0.1~dev.20240502.0729-1.fc40.x86_64/usr/share/graphviz/demo/*.php*
  error: File not found: /root/rpmbuild/BUILDROOT/graphviz-11.0.1~dev.20240502.0729-1.fc40.x86_64/usr/share/man/man3/*.3php.*
  warning: File not found: /root/rpmbuild/BUILDROOT/graphviz-11.0.1~dev.20240502.0729-1.fc40.x86_64/usr/lib64/graphviz/php/*.so

Gitlab: #2532, #2533, closes #2534