common closestSide: avoid truncating into 'int' when computing box sides (1/2)
a865bb66aef7e751f2c1a235819c77007b16050f converted this code to work on floating
point boxes (`boxf`) instead of integer boxes (`box`). However it did not update
the point type being used to store the result of the box calculations. The
effect of this was loss of precision, or even corruption when dealing with box
values that do not fit in an `int`.

This squashes 8 -Wfloat-conversion compiler warnings.

This is step 1 of a 2 step change, which the next commit will complete.