CMake: fix path to DevIL headers
A close reading of the FindDevIL.cmake docs¹ reveals that `IL_INCLUDE_DIR` is
actually the directory il.h and ilu.h live in, not the parent directory the IL/
directory lives in as one might expect from the convention in other CMake find
scripts. This change allows a build on macOS to find DevIL installed by
Homebrew. Presumably this caused no problems on Linux because the headers are
usually in the default system include path.

¹ https://cmake.org/cmake/help/latest/module/FindDevIL.html