CMake: fix: use full paths to Pango libraries when linking
It was previously observed (3ecf2f6206a486fecdea870b7f29ea539345642a,
f5868e0d90aebbeced03caaafe29498181ff53a5,
55104baf31ef53e5e78c861e359aaa2b0c2aeec7,
faa299532b29d8ec5eadb3c488fa855e7b0f3095) that macOS needs the full path to
PangoCairo libraries in order to locate them. It is unknown whether this is a
problem in their Homebrew installation formula, a problem in Graphviz’
FindPANGOCAIRO.cmake script, or a problem elsewhere in Graphviz’ CMake files. As
f5868e0d90aebbeced03caaafe29498181ff53a5 noted, there was suspicion that this
tweak was required everywhere. This seems to be confirmed when now trying to
enable the DevIL plugin on macOS which does not compile without this. Instead of
continuing to rediscover this problem once a year, lets fix it now globally.