smyrna binarySearch: simplify interaction with 'bsearch'
As discussed in the series merged in f050d1ddfb6016fddf562b9d9bd90d320b8e4a8a,
there is no requirement for the first argument to `bsearch` to have the same
type as the element type of the list being searched. This change takes advantage
of this to avoid constructing an extra temporary and having another layer of
pointer indirection.