add a test case for 'crow' arrowhead/tail placement when using ports
With Visual Studio Build Tools 2019, this test case failed deterministically on
Windows 10 using x64 Release builds and sporadically with Win32 Release builds,
but not with any Debug builds. An upgrade of Visual Studio Build Tools to 2022
made it pass for all build variants.

The reason for the problem with 2019 is not known, but seemed to have to do
with allocation of stack space or registers for local variables in the
`arrow_length_crow` function in arrows.c. Adding a simple `printf` statement
could make the problem disappear or appear, depending on at which line it was
placed.

GitLab: #2490