API BREAK: cgraph: replace node-by-id set with specialized implementation
The graph object contains a number of collections used for node lookup, one of
which is `n_id` indexed by node ID. Prior to this commit, its implementation was
a set using libcdt. libcdt implements generic collections and has served
Graphviz well for decades, with one of the lowest bug densities in the Graphviz
code base. However its design decisions were made at a time when hardware was
significantly different and, as a result, its performance characteristics are
not optimal on contemporary machines. In particular:

  1. libcdt is built around hierarchical tree structures. While this leads to a
     highly generic implementation, insertion and lookup require a lot of
     pointer chasing. On the deeply pipelined CPUs in contemporary machines,
     this is detrimental to performance. Each of these pointer dereferences
     presents to the hardware as an unpredictable indirect read, thwarting data
     prefetching (predictive read-ahead).

  2. The libcdt API leans heavily on function pointers in a way that lets
     callers customize numerous parts of its behavior. While this flexibility is
     a great feature in a generic data structure library, Graphviz only uses a
     narrow subset of behaviors. Profiling reveals calls through these function
     pointers often cannot be de-virtualized by the compiler, even when using
     link-time optimization. So each of these incur the runtime performance cost
     of a function call, as well as missed optimization opportunities that could
     occur from specializing the call targets.

  3. libcdt does dynamic memory allocation on common paths. For example, each
     element insertion incurs a call to `malloc`. Allocators were not fast when
     libcdt was designed, but perhaps their cost was subsumed by other factors.
     Contemporary heap allocators are reasonably fast, but it is still wise to
     avoid O(n) allocations.

This change reimplements this set, with an implementation specialized to its
use. The new code is a basic open addressed¹ hash table using linear probing.²
A node’s ID is used as its hash value, which is not good for diffusion³ but is
optimal for speed. Allocation calls are only made when expanding the table, thus
amortizing this cost.

The effect of this on the cycle count for processing the ftzz.gv graph from
#2475 is:

  ┌─────────┬─────────────┬─────────────┬────────┐
  │         │    before   │    after    │  diff  │
  ├─────────┼─────────────┼─────────────┼────────┤
  │ ftzz.gv │ 13592641273 │ 12263577203 │ -9.78% │
  └─────────┴─────────────┴─────────────┴────────┘

On the larger graph that ftzz.gv is representative of, it does not have as
dramatic an effect:

  ┌─────────┬────────┬────────┬────────┐
  │         │ before │ after  │  diff  │
  ├─────────┼────────┼────────┼────────┤
  │ bevy.gv │ 54m05s │ 52m54s │ -2.19% │
  └─────────┴────────┴────────┴────────┘

Gitlab: #2475

¹ https://en.wikipedia.org/wiki/Open_addressing
² https://en.wikipedia.org/wiki/Linear_probing
³ https://en.wikipedia.org/wiki/Confusion_and_diffusion#Diffusion