note ortho.c assertion failure fix
The accuracy improvements merged in dfd6aa4b9c54a408428392be09a733d22872b48b
seem to have unintentionally fixed a macOS assertion failure. The root cause of
this failure is not known, so it is possible this is only a coincidence and the
problem will reappear in future.

Gitlab: fixes #2538
Reported-by: Ryan Michela