cgraph: use a 'Agnode_t' instead of 'Agsubnode_t' when querying a node set
This removes an awkward wart where `Agsubnode_t *`s were passed into the
functions `node_set_find` and `node_set_remove`, but only the `->node` member of
the pointer needed to be valid. Note that this change unwraps some use of `AGID`
to avoid the casts inside `AGID` that otherwise trigger -Wcast-qual warnings.