tcldot graphcmd: rewrite 'write' to use public libgvc API
This code was calling private functions of libgvc. Surprisingly the only
platform to rigorously enforce this separation seems to be MinGW where
compilation fails:

  C:/msys64/mingw32/bin/../lib/gcc/i686-w64-mingw32/14.1.0/../../../../i686-w64-mingw32/bin/ld.exe: tclpkg/tcldot/CMakeFiles/tcldot.dir/tcldot-graphcmd.c.obj:tcldot-graphcm:(.text+0x1b2e): undefined reference to `gvjobs_output_langname'
  C:/msys64/mingw32/bin/../lib/gcc/i686-w64-mingw32/14.1.0/../../../../i686-w64-mingw32/bin/ld.exe: tclpkg/tcldot/CMakeFiles/tcldot.dir/tcldot-graphcmd.c.obj:tcldot-graphcm:(.text+0x1b54): undefined reference to `gvplugin_list'
  C:/msys64/mingw32/bin/../lib/gcc/i686-w64-mingw32/14.1.0/../../../../i686-w64-mingw32/bin/ld.exe: tclpkg/tcldot/CMakeFiles/tcldot.dir/tcldot-graphcmd.c.obj:tcldot-graphcm:(.text+0x1bc4): undefined reference to `gvrender_select'

Layout is now done unconditionally (even if `LAYOUT_NOT_REQUIRED` is set). There
does not seem to be a way to detect this using the public API, but this should
be a minor performance concern for most users. The core issue is that the GVC
interface basically looks like this:

  layout() {
    engine = <what is -K…?>
    job->flags |= engine->flags
    <do layout>
  }

  render() {
    if (<layout not done>) {
      return error
    }
    engine = <what is -T…?>
    job->flags |= engine->flags
    <do render>
  }

So while it appears the caller has access to everything they need in
`job->flags`, they cannot actually determine whether the renderer they will
target has `LAYOUT_NOT_REQUIRED` set without attempting rendering, which in turn
aborts it layout has not been done.

Gitlab: #2337, #2530, fixes #2546