tclpkg tcldot: pass both interpreter and canvas through GVC job context
The TCL bindings were abusing the `imagedata` member of the GVC job – which is
intended to contain pixel values of a rendered image – to pass an extra piece of
context to the TK renderer. To migrate the TCL bindings to use the public GVC,
we need to stop doing this and pass context exclusively through the `context`
member.

Gitlab: #2530, #2546