dotgen: hoist crossing count array out of inner loop
When performing ranking calculations, the `Count` array was being allocated and
freed each time through `rcross`. While this is functionally correct – a call to
this function does not need the values from the previous call – `rcross`
effectively sits within a nested loop, one that is particularly hot within the
ftzz.gv graph from #2475. Repeatedly allocating and freeing these arrays turn
out to play a significant role in the overall cost of processing this graph.

This change (1) hoists the array outwards and (2) makes it lazily extend. That
is, the allocation and free of the array is moved outwards through callers. And
the (now long lived) array is only expanded when we need to add a non-zero value
to one of its entries. The array is effectively a sparsely populated infinite
list, with values beyond the concrete end treated as zero. This change has the
following effect on the cycle count of processing ftzz.gv:

  ┌─────────┬─────────────┬─────────────┬─────────┐
  │         │    before   │    after    │   diff  │
  ├─────────┼─────────────┼─────────────┼─────────┤
  │ ftzz.gv │ 12266216639 │ 10577470113 │ -13.78% │
  └─────────┴─────────────┴─────────────┴─────────┘

Meanwhile it has negligible effect on bevy.gv, the larger graph ftzz.gv is
intended to be representative of. So it seems we will need to profile bevy.gv
itself to get any future wins there.

In theory, a compiler could make this change on its own. However, GCC 11.4.0
does not, even at -O3 with link-time optimization (LTO). This is probably not
surprising as (1) this would be a fairly sophisticated compiler optimization and
(2) it probably requires profile-guided optimization (PGO) for the compiler to
realize that such a transformation would be valuable.

Gitlab: #2475
Reported-by: Alex Saveau